requirements.txt
tkinter
pyautogui
pillow
keyboard